+++ bin/pulsar.dart
@@ -7,7 +7,7 @@ import 'package:pulsar_cli/src/commands/doctor_command.dart';
 import 'package:pulsar_cli/src/commands/serve_command.dart';
 
 Future<void> main(List<String> args) async {
-  final String cliVersion = '0.2.7';
+  final String cliVersion = '0.2.9';
   final CommandRunner runner =
       CommandRunner(
           'pulsar',