+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.9
+- Avoided FOUC and added support for the latest version of Pulsar 0.4.7
+- CSS minification on release mode build
+
 ## 0.2.8
 - Quick fix for create command prompt